chore(acp): split terminal output rendering
