+++ docs/journal/2026-04-05-acp-conversation-tool-content-split.md
@@ -6,6 +6,7 @@
 - extracted `request_user_input` pending/result cards into `web/src/components/acp_request_user_input_cards.tsx`
 - extracted tool-call bubble/group/explore rendering plus fold/live-state helpers into `web/src/components/acp_tool_bubbles.tsx`
 - split payload/text/diff rendering a second time into `web/src/components/acp_tool_payload_content.tsx` so `acp_tool_content.tsx` can focus on terminal rendering, ANSI caches, and payload normalization helpers
+- moved terminal rendering plus ANSI cache accounting into `web/src/components/acp_terminal_output.tsx` so `acp_tool_content.tsx` is now a small aggregation layer
 - kept `acp_conversation.tsx` focused on conversation item dispatch, cache wiring, key/focus helpers, and scroll behavior
 - preserved the existing public cache helpers by re-exporting `parseAnsiSegmentsCached` from `acp_conversation.tsx`
 - preserved the existing public live/fold helper exports by re-exporting them from `acp_conversation.tsx`
